Separate headers and footers so that they can be reused
