+++ date-contact.php
@@ -6,7 +6,7 @@ if(defined('DEBUG')) {
 
 try {
   require_once __DIR__ . '/vendor/autoload.php';
-  require_once __DIR__ . '/config.php';
+  require_once __DIR__ . '/include/config.php';
 
   $client = new Google_Client();
