+++ index.php
@@ -1,57 +1,8 @@
 <?php
-require 'functions.php';
+$page = "acasa";
 
-$data = get_data();
+require 'include/header.php';
 ?>
-<!DOCTYPE html>
-<html lang="ro">
-  <head>
-    <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <meta name="description" content="Inițiativa cetățenească de modificare a Constituției">
-    <meta name="author" content="USR">
-    <meta property="og:url" content="<?= SITE_URL ?>">
-    <meta property="og:title" content="Fără penali în funcții publice">
-    <meta property="og:description" content="Inițiativa cetățenească de modificare a Constituției">
-    <meta property="og:image" content="<?= SITE_URL ?>build/img/facebook.png">
-    <link href="build/img/favicon.ico" rel="apple-touch-icon">
-    <link href="build/img/favicon.ico" rel="icon">
-    <title>Fără penali în funcții publice &middot; Inițiativa cetățenească de modificare a Constituției</title>
-    <link href="https://code.cdn.mozilla.net/fonts/fira.css" rel="stylesheet">
-    <link href="build/css/style.min.css" rel="stylesheet">
-    <!--[if lt IE 9]>
-      <script src="https://oss.maxcdn.com/html5shiv/3.7.3/html5shiv.min.js"></script>
-      <script src="https://oss.maxcdn.com/respond/1.4.2/respond.min.js"></script>
-    <![endif]-->
-  </head>
-  <body>
-  <nav class="navbar navbar-default">
-    <div class="container">
-      <div class="navbar-header">
-        <button type="button" class="navbar-toggle collapsed" data-toggle="collapse" data-target="#bs-example-navbar-collapse-1" aria-expanded="false">
-          <span class="sr-only">Toggle navigation</span>
-          <span class="icon-bar"></span>
-          <span class="icon-bar"></span>
-          <span class="icon-bar"></span>
-        </button>
-        <a class="navbar-brand" href="<?= SITE_URL ?>">
-          <img src="build/img/logo.png" alt="USR">
-        </a>
-      </div>
-      <div class="collapse navbar-collapse" id="bs-example-navbar-collapse-1">
-        <ul class="nav navbar-nav navbar-right">
-          <li><a href="#despre">Despre</a></li>
-          <?php if (CAMPANIE_DE_SEMNATURI) { ?>
-            <li><a href="#semneaza">Semnează inițativa</a></li>
-            <li><a href="#harta">Situația pe regiuni</a></li>
-          <?php } ?>
-          <li><a href="#media">Media</a></li>
-          <li><a href="#comunicate">Comunicate de presă</a></li>
-        </ul>
-      </div>
-    </div>
-  </nav>
     <div class="jumbotron jumbotron-primary">
       <div class="container">
         <div class="row">
@@ -303,7 +254,7 @@ $data = get_data();
         </ul>
         <div class="col-md-9">
           <div class="tab-content">
-            <div class="tab-pane fade show active in" id="media-tabs-video" role="tabpanel" aria-labelledby="media-tabs-video-tab">
+            <div class="tab-pane fade active in" id="media-tabs-video" role="tabpanel" aria-labelledby="media-tabs-video-tab">
                 <div class="carousel slide" id="carousel-videos" data-ride="carousel" data-interval="false">
                   <div class="carousel-inner" role="listbox">
                     <div class="item active embed-responsive embed-responsive-16by9">
@@ -360,94 +311,6 @@ $data = get_data();
         </div>
       </div>
     </div>
-    <div class="implica-te">
-      <div class="container">
-          <h1>Implică-te și tu!</h1>
-          <a href="#formular" class="btn btn-inverted-dark-bg">Alătură-te inițiativei</a>
-      </div>
-    </div>
-    <footer class="container">
-      <div class="col-md-3 contact">
-          <h4>Contact</h4>
-          <div>Bulevardul Aviatorilor Nr. 9,<br>Sector 1, Bucureşti</div>
-      </div>
-      <div class="col-md-2 col-md-push-2 navigare">
-          <h4>Navigare</h4>
-          <ul>
-            <li><a href="#despre">Despre</a></li>
-            <li><a href="#semneaza">Semnează inițativa</a></li>
-            <li><a href="#harta">Situația pe regiuni</a></li>
-            <li><a href="#media">Media</a></li>
-            <li><a href="#comunicate">Comunicate de presă</a></li>
-          </ul>
-      </div>
-      <div class="col-md-3 col-md-push-4 distribuie">
-          <h4>Distribuie</h4>
-          <a class="btn btn-facebook" href="https://www.facebook.com/sharer.php?u=<?= SITE_URL ?>" target="_blank">
-            <img src="build/img/logo-facebook-inverted.png" alt="Facebook">
-            Distribuie
-          </a>
-      </div>
-      <div class="col-md-12 final">
-          <div class="usr">
-            Uniunea Salvați România &copy; 2018
-          </div>
-          <div class="rezistenta">
-            În parteneriat cu <a href="https://www.rezist.ro/" title="Rezistența"><img src="build/img/rezistenta.png" alt="Rezistența"></a>
-          </div>
-      </div>
-    </footer>
-    <div class="social-media">
-      <a class="facebook" href="https://www.facebook.com/sharer.php?u=<?= SITE_URL ?>" target="_blank">
-        <img src="build/img/logo-facebook-inverted.png" alt="Facebook">
-        Distribuie
-      </a>
-    </div>
 <?php
-function modal_locatii($id, $title, $data_modal) {
-  global $data;
+require_once 'include/footer.php';
 ?>
-  <div class="modal fade" id="<?= $id ?>" tabindex="-1" role="dialog" aria-labelledby="<?= $id ?>-label">
-    <div class="modal-dialog modal-lg" role="document">
-      <div class="modal-content">
-        <div class="modal-header">
-          <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>
-          <h4 class="modal-title" id="<?= $id ?>-label"><?= $title ?></h4>
-        </div>
-        <div class="modal-body">
-          <div class="row row-three-columns">
-            <?php foreach ($data_modal as $judet => $locatii) { ?>
-              <div class="col-sm-6 col-md-4">
-                <h4><?= $data['numeJudet'][$judet] ?></h4>
-                <?php foreach ($locatii as $locatie) { ?>
-                  <dl>
-                    <dt><?= $locatie['locatie'] ?></dt>
-                    <dd>
-                      <ul class="list-unstyled">
-                        <?php foreach ($locatie['intrari'] as $intrare) { ?>
-                          <li><?= $intrare ?>
-                        <?php } ?>
-                      </ul>
-                    </dd>
-                  </dl>
-                <?php } ?>
-              </div>
-            <?php } ?>
-          </div>
-        </div>
-        <div class="modal-footer">
-          <button type="button" class="btn btn-primary" data-dismiss="modal">Închide</button>
-        </div>
-      </div>
-    </div>
-  </div>
-<?php
-}
-
-modal_locatii('contact', 'Persoane de contact', $data['contacte']);
-modal_locatii('corturi', 'Corturile de campanie ale USR', $data['corturi']);
-?>
-    <script><?php printf('remoteData = %s;', json_encode($data)); ?></script>
-    <script src="build/js/application.min.js"></script>
-  </body>
-</html>
