+++ src/less/jumbotron.less
@@ -20,16 +20,6 @@
     padding-right: 0;
   }
 
-  @media (max-width: @screen-md) {
-    margin-bottom: 70px;
-  }
-
-  @media (max-width: @screen-sm-max) {
-    margin-bottom: 120px;
-  }
-
-  margin-bottom: 100px;
-
   .banner {
     margin-bottom: 0;
 
@@ -192,4 +182,16 @@
   .countdown-separator {
     vertical-align: top;
   }
+}
+
+.page-acasa .jumbotron-primary {
+  @media (max-width: @screen-md) {
+    margin-bottom: 70px;
+  }
+
+  @media (max-width: @screen-sm-max) {
+    margin-bottom: 120px;
+  }
+
+  margin-bottom: 100px;
 }
