+++ src/less/site.less
@@ -12,7 +12,7 @@ input {
   }
 }
 
-.comunicate, .descriere {
+.comunicate, .descriere, .page-grupul-de-initiativa .formular {
   background-color: #d7e6f2;
   font-size: 14px;
 }
