+++ src/less/style.less
@@ -18,3 +18,4 @@
 
 // Global styling
 @import "site.less";
+@import "grupul-de-initiativa.less";
